Final discussion changes AB#15164
